signing-proxy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
alibabacloud-openapi-util==0.2.4
Tea==0.1.7
